GameMaster split: GameMaster: method that was not relevant to play a game were excluded to New class file GameMechanic: created. GameMechanic: New class, its purpose is to be medium between Interfejs and Plansza. file PositionGenerator: created. PositionGenerator: New class, its purpose is to generate every possible position from existing one and return it. file InterfejsGeneratora: created. InterfejsGeneratora: New class, its purpose is to be PositionGenerators aid to play moves on the board. Interfejs: Interfejs now extends ABC.
