+++ Util/Information.py
@@ -1,3 +1,4 @@
 
 directionOptions = ["WN", "W", "WS", "ES", "E", "EN"]
 resourceFieldCoordinates = [[1, 0, -1], [-2, 3, -1], [1, -3, 2]]
+board_size = 4
